+++ Expanded Cheat Menu/Files/Code/OnMsgs.lua
@@ -867,21 +867,6 @@ do -- LoadGame/CityStart
     if not UICity then
       return
     end
-PlaceObj('Sound', {
-	'name', "Object PreciousExtractor Select",
-	'type', "ObjectOperation",
-	'volume', 150,
-	'mindistance', 2500,
-}, {
-	PlaceObj('Sample', {
-		'file', "Sounds/Objects/ExtractorPrecious/extractorUniversal_select1",
-		'frequency', 50,
-	}),
-	PlaceObj('Sample', {
-		'file', "Sounds/Objects/ExtractorPrecious/extractorUniversal_select2",
-		'frequency', 50,
-	}),
-	})
 
     -- a place to store per-game values
     if not UICity.ChoGGi then
@@ -1088,6 +1073,21 @@ PlaceObj('Sound', {
 
 
 
+    -- it seems the devs didn't remove all traces of the old select sound, or didn't add them all?
+    if LuaRevision > 233467 then
+      RebuildFXRules()
+      local tab = UICity.labels.PreciousMetalsExtractor or ""
+      for i = 1, #tab do
+        ChoGGi.CodeFuncs.ToggleWorking(tab[i])
+      end
+    elseif FXRules.SelectObj.start.PreciousMetalsExtractor.any[1].Source == "Actor" then
+      table.remove(FXRules.SelectObj.start.PreciousMetalsExtractor.any,1)
+      RemoveFromRules("Object PreciousExtractor Select")
+      local tab = UICity.labels.PreciousMetalsExtractor or ""
+      for i = 1, #tab do
+        ChoGGi.CodeFuncs.ToggleWorking(tab[i])
+      end
+    end
 
     -- bloody hint popups
     if ChoGGi.UserSettings.DisableHints then